refactor(services): move publication state toggle to a single service
