--- server/services/publish-service.js
@@ -1,30 +1,0 @@
-'use strict';
-
-const { getPluginEntityUid } = require('../utils/getEntityUId');
-
-const actionUId = getPluginEntityUid('action');
-
-module.exports = ({ strapi }) => ({
-	/**
-	 * Publish a single record
-	 *
-	 */
-	async index(record) {
-		// handle single content type, id is always 1
-		const entityId = record.entityId || 1;
-
-		const entity = await strapi.entityService.findOne(record.entitySlug, entityId);
-
-		// ensure entity has not been publish yet
-		if (!entity.publishedAt) {
-			await strapi.entityService.update(record.entitySlug, entityId, {
-				data: {
-					publishedAt: new Date(),
-				},
-			});
-		}
-
-		// remove any used actions
-		strapi.entityService.delete(actionUId, record.id);
-	},
-});
